Re-arrange files and factor out controller utils
