+++ lib/controllers/features.js
@@ -9,82 +9,15 @@ var Promise = require('bluebird');
 
 var CacheItem = require('../models/CacheItem');
 
-var layer = require('../layer.js');
 var metrics = require('../metrics/pipeline');
-var featuresDS = require('../ds-features');
+var layer = require('../layer.js');
+var featuresDS = require('../datasources/features');
+var settings = require('../settings.js');
+var util = require('./util.js');
 
-var DEFAULT_BUFFER = 0.05;
 
 Promise.promisifyAll(CacheItem);
 
-function bufferStream(stream, done) {
-  var bufs = [];
-  var length = 0;
-
-  stream.on('readable', function () {
-    var buf = stream.read();
-    bufs.push(buf);
-    length += buf.length;
-  });
-
-  stream.on('end', function () {
-    done(null, Buffer.concat(bufs, length));
-  });
-
-  stream.on('error', function (error) {
-    done(error);
-  });
-}
-
-function createRenderStream(map, tile) {
-  var passThrough = new stream.PassThrough();
-  var bounds = nodetiles.projector.util.tileToMeters(tile[1], tile[2], tile[0]);
-
-  // Start the timer.
-  var stopMetric = metrics.renderTimer();
-
-  map.render({
-    bounds: {minX: bounds[0], minY: bounds[1], maxX: bounds[2], maxY: bounds[3]},
-    width: 256,
-    height: 256,
-    zoom: tile[0],
-    callback: function (error, canvas) {
-      if (error) {
-        passThrough.emit('error', error);
-        return;
-      }
-
-      var stream = canvas.createPNGStream();
-
-      // Stop the timer when the stream ends.
-      stream.on('end', stopMetric);
-
-      // TODO: handle the error
-      stream.pipe(passThrough);
-    }
-  });
-
-  return passThrough;
-}
-
-function createGrid(map, tile, options, done) {
-  if (done === undefined) {
-    done = options;
-    options = {};
-  }
-
-  var drawImage = (options.drawImage === true);
-
-  var bounds = nodetiles.projector.util.tileToMeters(tile[1], tile[2], tile[0]);
-  map.renderGrid({
-    bounds: {minX: bounds[0], minY: bounds[1], maxX: bounds[2], maxY: bounds[3]},
-    width: 64,
-    height: 64,
-    zoom: tile[0],
-    drawImage: drawImage,
-    callback: done
-  });
-}
 
 // Keep track of the different surveys we have maps for
 // TODO: use a fixed-size LRU cache, so this doesn't grow without bounds.
@@ -101,7 +34,7 @@ function getMap(options, callback) {
 
   var boundsBuffer = options.boundsBuffer;
   if (boundsBuffer === undefined) {
-    boundsBuffer = DEFAULT_BUFFER;
+    boundsBuffer = settings.defaultBuffer;
   }
 
   // Set up the map
@@ -156,7 +89,7 @@ var world = nodetiles.projector.util.tileToMeters(0, 0, 0);
 
 function smartBuffer(bounds, buffer) {
   if (buffer === undefined) {
-    buffer = DEFAULT_BUFFER;
+    buffer = settings.defaultBuffer;
   }
 
   var amountX = (bounds.maxX - bounds.minX) * buffer;
@@ -225,7 +158,7 @@ exports.render = function render(req, res, next) {
     options.select = layerDef.select;
     options.styles = layerDef.styles;
     return getMap(options).then(function (map) {
-      bufferStream(createRenderStream(map, tile), handleData);
+      util.bufferStream(util.createRenderStream(map, tile), handleData);
     });
   }).catch(function (error) {
     console.log('Error generating tile', error);
@@ -273,16 +206,16 @@ exports.renderGrids = function renderGrids(req, res, next) {
     return getMap(options)
     .then(function (map) {
       if (format === 'png') {
-        return Promise.promisify(createGrid)(map, tile, { drawImage: true })
+        return Promise.promisify(util.createGrid)(map, tile, { drawImage: true })
         .then(function (canvas) {
           var passThrough = new stream.PassThrough();
           canvas.createPNGStream().pipe(passThrough);
-          bufferStream(passThrough, handleData);
+          util.bufferStream(passThrough, handleData);
         });
       }
 
       var stopTimer = metrics.gridTimer();
-      return Promise.promisify(createGrid)(map, tile)
+      return Promise.promisify(util.createGrid)(map, tile)
       .finally(function () {
         // Stop the metric timer whether or not this succeeded.
         stopTimer();
