+++ lib/controllers/tiles.js
@@ -11,9 +11,9 @@ var CacheItem = require('../models/CacheItem');
 
 var layer = require('../layer.js');
 var metrics = require('../metrics/pipeline');
-var responsesDS = require('../ds-responses');
-
-var DEFAULT_BUFFER = 0.05;
+var responsesDS = require('../datasources/responses');
+var settings = require('../settings.js');
+var util = require('./util.js');
 
 Promise.promisifyAll(CacheItem);
 
@@ -102,7 +102,7 @@ function getMap(options, callback) {
 
   var boundsBuffer = options.boundsBuffer;
   if (boundsBuffer === undefined) {
-    boundsBuffer = DEFAULT_BUFFER;
+    boundsBuffer = settings.defaultBuffer;
   }
 
   // Set up the map
@@ -146,7 +146,7 @@ function getMap(options, callback) {
     projection: 'EPSG:4326',
     query: query,
     select: select,
-    grid: grid,
+    grid: grid
   });
 
   // Wrap the datasource.getShapes call so we can measure its duration
@@ -173,44 +173,7 @@ function getMap(options, callback) {
   return Promise.resolve(map).nodeify(callback);
 }
 
-var world = nodetiles.projector.util.tileToMeters(0, 0, 0);
 
-function smartBuffer(bounds, buffer) {
-  if (buffer === undefined) {
-    buffer = DEFAULT_BUFFER;
-  }
-
-  var amountX = (bounds.maxX - bounds.minX) * buffer;
-  var amountY = (bounds.maxY - bounds.minY) * buffer;
-
-  var minX = bounds.minX - amountX;
-  var maxX = bounds.maxX + amountX;
-  var minY = bounds.minY - amountY;
-  var maxY = bounds.maxY + amountY;
-
-  if (minX < world[0]) {
-    minX = world[0];
-  }
-
-  if (maxX > world[2]) {
-    maxX = world[2];
-  }
-
-  if (minY < world[1]) {
-    minY = world[1];
-  }
-
-  if (maxY > world[3]) {
-    maxY = world[3];
-  }
-
-  return {
-    minX: minX,
-    minY: minY,
-    maxX: maxX,
-    maxY: maxY
-  };
-}
 
 /**
  * Render a tile using a map that we create or an existing cached map.
@@ -228,7 +191,7 @@ exports.render = function render(req, res, next) {
   res.set('Content-Type', 'image/png');
 
   var options = {
-    boundsBuffer: smartBuffer,
+    boundsBuffer: util.smartBuffer,
     surveyId: surveyId
   };
 
@@ -292,7 +255,7 @@ exports.renderGrids = function renderGrids(req, res, next) {
 
   // We'll use these options to create the map
   var options = {
-    boundsBuffer: smartBuffer,
+    boundsBuffer: util.smartBuffer,
     surveyId: surveyId,
     type: 'grid'
   };
