+++ lib/datasources/responses.js
@@ -4,20 +4,15 @@ var stream = require('stream');
 var util = require('util');
 
 var _ = require('lodash');
-var knex = require('knex');
 var projector = require('nodetiles-core').projector;
 
-var metrics = require('./metrics/pipeline');
-var settings = require('./settings');
+var metrics = require('../metrics/pipeline');
+var Response = require('../models/Response');
+var settings = require('../settings');
 
 var maxResponseCount = settings.maxResponseCount;
 
 var FEATURE = 'Feature';
-
-var pg = require('knex')({
-  client: 'pg',
-  connection: settings.psqlConnectionString
-});
 
 function ToFeatures(options) {
   // Start and pause the timer here, in case we encounter 0 documents in the
@@ -32,20 +27,22 @@ function ToFeatures(options) {
   this.procTimer.pause();
 }
 
-util.inherits(ToFeatures, stream.Transform);
-
 ToFeatures.prototype._transform = function transform(item, encoding, done) {
   this.procTimer.start();
 
   item.type = FEATURE;
 
-  // TODO
-  // We can probably find a much faster way to do this than using omit.
+
   if (!item.properties) {
-    item.properties = _.omit(item, 'geometry');
+    item.properties = {};
   }
 
-  item.geometry = JSON.parse(item.geometry);
+  // Copy all the relevant data items to the properties field.
+  var entries = item.entries;
+  if (entries) {
+    item.entries = undefined;
+    _.assign(item.properties, entries[entries.length - 1]);
+  }
 
   if (this.project) {
     item = this.project(item);
@@ -56,6 +53,7 @@ ToFeatures.prototype._transform = function transform(item, encoding, done) {
   done();
 };
 
+util.inherits(ToFeatures, stream.Transform);
 
 
 exports.create = function create(options) {
@@ -65,11 +63,10 @@ exports.create = function create(options) {
   var grid = options.grid;
 
   function getShapes(minX, minY, maxX, maxY, mapProjection, done) {
-    var finish = _.once(done);
-
-    // Set up the coordinates we need.
     var sw = [minX, minY];
     var ne = [maxX, maxY];
+
+    var finish = _.once(done);
 
     // project request coordinates into data coordinates
     if (mapProjection !== projection) {
@@ -77,11 +74,22 @@ exports.create = function create(options) {
       ne = projector.project.Point(mapProjection, projection, ne);
     }
 
-    //  Make them more readable later
+    var query = _.clone(baseQuery);
+
     var west = sw[0];
     var south = sw[1];
     var east = ne[0];
     var north = ne[1];
+    var boundingCoordinates = [ [ [west, south], [west, north], [east, north], [east, south], [west, south] ] ];
+
+    query.indexedGeometry = {
+      $geoIntersects: {
+        $geometry: {
+          type: 'Polygon',
+          coordinates: boundingCoordinates
+        }
+      }
+    };
 
     var dbTimer = metrics.dbTimer();
     var procTimer = metrics.processingTimer();
@@ -93,41 +101,45 @@ exports.create = function create(options) {
       };
     }
 
-    var k = pg.select(knex.raw('ST_AsGeoJson(geom) as geometry, id, source, type, object_id, short_name, long_name, info, timestamp'))
-              // We use knex.raw because ST_AsGeoJson gets interpreted wrong
-              // without it.
-              // TODO: Use the "select" value to only get the fields we need
-              .from(settings.featuresTable)
-              .where(knex.raw('geom && ST_MakeEnvelope(?, ?, ?, ?)', [west, south, east, north]))
-              .andWhere(baseQuery);
-
-    // console.log(k.toString()); // Debug the query
-
-    dbTimer.start();
-    var docStream = k.stream();
-    var features = [];
-    var len = features.length;
-
     var toFeatures = new ToFeatures({
       grid: grid,
       project: project,
       procTimer: procTimer
     });
 
-    var featureStream = docStream.pipe(toFeatures)
-        .on('readable', function () {
-          features[len] = featureStream.read();
-          len += 1;
-        })
-        .on('end', function () {
-          dbTimer.stop();
-          procTimer.stop();
+    dbTimer.start();
+    var docStream = Response.find(query)
+    .select(select)
+    .lean()
+    .limit(maxResponseCount)
+    .stream()
+    .on('error', function (error) {
+      finish(error);
+    })
+    .on('close', function () {
+      dbTimer.pause();
+    });
 
-          finish(null, {
-            type: 'FeatureCollection',
-            features: features
-          });
-        });
+    var features = [];
+    var len = features.length;
+
+    var featureStream = docStream.pipe(toFeatures)
+    .on('readable', function () {
+      features[len] = featureStream.read();
+      len += 1;
+    })
+    .on('end', function () {
+      dbTimer.stop();
+      procTimer.stop();
+
+      finish(null, {
+        type: 'FeatureCollection',
+        features: features
+      });
+    })
+    .on('error', function (error) {
+      finish(error);
+    });
   }
 
   return {
