+++ lib/routes.js
@@ -45,11 +45,11 @@ if (settings.cacheMethod === 'mongo') {
   });
 }
 
-console.log("Using cache", settings.cacheMethod);
 
 if (settings.nocache) {
   useEtagCache = noOp;
 }
+
 
 // Parse the tilename parameters from the URL and store as a res.locals.tile
 // array.
@@ -95,6 +95,63 @@ function makeQueryDigester(param, field) {
 }
 
 exports.setup = function setup(app) {
+  // Feature tile routes -------------------------------------------------------
+
+  // Create layer definition
+  app.post(
+    '/features/tile.json',
+    makePostDigester('layerId'),
+    bodyParser.json(),
+    tileJson.post
+  );
+
+  app.get(
+    '/features/tile.json',
+    makeQueryDigester('layerDefinition', 'layerId'),
+    tileJson.get
+  );
+
+  app.get('/features/layers/:layerId', tileJson.getLayerDef);
+
+  app.get(
+    '/features/layers/:layerId/tiles/:zoom/:x/:y.png',
+    parseTileName,
+    // useEtagCache,
+    useCacheFeatures,
+    features.render
+  );
+
+  app.get(
+    '/features/layers/:layerId/utfgrids/:zoom/:x/:y.:format',
+    parseTileName,
+    // useEtagCache,
+    useCacheFeatures,
+    features.renderGrids
+  );
+
+  // Getting features without a layer definition
+
+  // Get a PNG tile for features
+  app.get(
+    '/features/tiles/:zoom/:x/:y.png',
+    parseTileName,
+    // useEtagCache,
+    // useCache,
+    features.render
+  );
+
+  // Get a grid tile for features
+  app.get(
+    '/features/utfgrids/:zoom/:x/:y.:format',
+    parseTileName,
+    // useEtagCache,
+    // useCache,
+    features.renderGrids
+  );
+
+
+  // Survey tile routes --------------------------------------------------------
+
   // Create layer definition/get tile.json
   app.post(
     '/surveys/:surveyId/tile.json',
@@ -102,6 +159,7 @@ exports.setup = function setup(app) {
     bodyParser.json(),
     tileJson.post
   );
+
   app.get(
     '/surveys/:surveyId/tile.json',
     makeQueryDigester('layerDefinition', 'layerId'),
@@ -130,66 +188,7 @@ exports.setup = function setup(app) {
   );
 
 
-  /* Features --------------------------------------------------------------- */
-  app.post(
-    '/features/tile.json',
-    makePostDigester('layerId'),
-    bodyParser.json(),
-    tileJson.post
-  );
-
-  app.get(
-    '/features/tile.json',
-    makeQueryDigester('layerDefinition', 'layerId'),
-    tileJson.get
-  );
-
-  // Get a layer definition
-  app.get('/surveys/:surveyId/layers/:layerId', tileJson.getLayerDef);
-
-  app.get(
-    '/features/layers/:layerId/tiles/:zoom/:x/:y.png',
-    parseTileName,
-    // useEtagCache,
-    useCacheFeatures,
-    features.render
-  );
-
-  app.get(
-    '/features/layers/:layerId/utfgrids/:zoom/:x/:y.:format',
-    parseTileName,
-    // useEtagCache,
-    useCacheFeatures,
-    features.renderGrids
-  );
-
-  /* Features w/out a layer def ============================================= */
-
-  // Get a PNG tile for features
-  app.get(
-    '/features/tiles/:zoom/:x/:y.png',
-    parseTileName,
-    // useEtagCache,
-    // useCache,
-    features.render
-  );
-
-  // Get a grid tile for features
-  app.get(
-    '/features/utfgrids/:zoom/:x/:y.:format',
-    parseTileName,
-    // useEtagCache,
-    // useCache,
-    features.renderGrids
-  );
-
-
-
-
-
-  /**
-   * Legacy routes
-   */
+  // Legacy survey tile routes -------------------------------------------------
 
   // Get a tile for a survey
   app.get('/:surveyId/tiles/:zoom/:x/:y.png', parseTileName, useEtagCache, useCache, tiles.render);
