+++ lib/s3-cache-features.js
@@ -127,9 +127,6 @@ module.exports = function setup(options) {
 
       // Cache the file using S3
       if (s3client) {
-
-        console.log("Checking response content type", res.get('Content-Type'));
-
         saveMetric = metrics.s3Storage();
         var name = makeTileName(req);
         var headers = {
