+++ lib/settings.js
@@ -2,6 +2,10 @@
 
 var settings = module.exports;
 
+// Default buffer added around tile queries
+// You probably won't need to override this default.
+settings.defaultBuffer = process.env.DEFAULT_BUFFER || 0.05;
+
 // S3 parameters for storing generated tiles
 settings.s3Key = process.env.S3_KEY;
 settings.s3Secret = process.env.S3_SECRET;
